requirements.txt
python-dotenv>=0.19.0
requests>=2.25.0
kafka-python>=2.0.2
pyspark>=3.1.2
minio>=7.0.0
elasticsearch>=7.13.0,<8.0.0
python-telegram-bot>=13.7,<14.0
apache-airflow==2.2.5 